move isolated changes to 0.12

+++ packages/core/src/indexing-store/realtime.ts
@@ -2,6 +2,7 @@ import type { QB } from "@/database/queryBuilder.js";
 import type { Common } from "@/internal/common.js";
 import {
   DbConnectionError,
+  InvalidStoreMethodError,
   RawSqlError,
   RecordNotFoundError,
   RetryableError,
@@ -15,6 +16,7 @@ import { drizzle } from "drizzle-orm/pg-proxy";
 import {
   type IndexingStore,
   checkOnchainTable,
+  checkTableAccess,
   validateUpdateSet,
 } from "./index.js";
 import { getCacheKey, getWhereCondition } from "./utils.js";
@@ -23,10 +25,12 @@ export const createRealtimeIndexingStore = ({
   common,
   schemaBuild: { schema },
   indexingErrorHandler,
+  chainId,
 }: {
   common: Common;
   schemaBuild: Pick<SchemaBuild, "schema">;
   indexingErrorHandler: IndexingErrorHandler;
+  chainId?: number;
 }): IndexingStore => {
   let qb: QB = undefined!;
 
@@ -62,6 +66,7 @@ export const createRealtimeIndexingStore = ({
         method: "find",
       });
       checkOnchainTable(table, "find");
+      checkTableAccess(table, "find", key, chainId);
       return find(table, key);
     }),
     // @ts-ignore
@@ -103,6 +108,13 @@ export const createRealtimeIndexingStore = ({
                 return rows;
               };
 
+              if (chainId !== undefined)
+                Array.isArray(values)
+                  ? values.map((v) =>
+                      checkTableAccess(table, "insert", v, chainId),
+                    )
+                  : checkTableAccess(table, "insert", values, chainId);
+
               return qb.wrap((db) =>
                 db
                   .insert(table)
@@ -122,6 +134,7 @@ export const createRealtimeIndexingStore = ({
               if (Array.isArray(values)) {
                 const rows = [];
                 for (const value of values) {
+                  checkTableAccess(table, "insert", value, chainId);
                   const row = await find(table, value);
 
                   if (row) {
@@ -154,6 +167,7 @@ export const createRealtimeIndexingStore = ({
                 }
                 return rows;
               } else {
+                checkTableAccess(table, "insert", values, chainId);
                 const row = await find(table, values);
 
                 if (row) {
@@ -189,6 +203,12 @@ export const createRealtimeIndexingStore = ({
                   method: "insert",
                 });
                 checkOnchainTable(table, "insert");
+                if (chainId !== undefined)
+                  Array.isArray(values)
+                    ? values.map((v) =>
+                        checkTableAccess(table, "insert", v, chainId),
+                      )
+                    : checkTableAccess(table, "insert", values, chainId);
 
                 // Note: `onConflictDoNothing` is used to ensure the transaction will only fail
                 // for connection issues. This is a workaround to avoid the transaction being aborted.
@@ -221,8 +241,9 @@ export const createRealtimeIndexingStore = ({
                     .then(parseResult),
                 );
               })().then(onFulfilled, onRejected),
-            catch: (onRejected) => inner.then(undefined, onRejected),
-            finally: (onFinally) =>
+            catch: (onRejected): Promise<any> =>
+              inner.then(undefined, onRejected),
+            finally: (onFinally): Promise<any> =>
               inner.then(
                 (value: any) => {
                   onFinally?.();
@@ -249,6 +270,7 @@ export const createRealtimeIndexingStore = ({
             method: "update",
           });
           checkOnchainTable(table, "update");
+          checkTableAccess(table, "update", key, chainId);
 
           const row = await find(table, key);
           if (typeof values === "function") {
@@ -290,6 +312,7 @@ export const createRealtimeIndexingStore = ({
         method: "delete",
       });
       checkOnchainTable(table, "delete");
+      checkTableAccess(table, "delete", key, chainId);
 
       const deleted = await qb.wrap((db) =>
         db.delete(table).where(getWhereCondition(table, key)).returning(),
@@ -300,6 +323,10 @@ export const createRealtimeIndexingStore = ({
     // @ts-ignore
     sql: drizzle(
       errorHandler(async (_sql, params, method, typings) => {
+        if (chainId !== undefined)
+          throw new InvalidStoreMethodError(
+            `Raw SQL queries are not allowed in 'isolated' ordering.`,
+          );
         const query: QueryWithTypings = { sql: _sql, params, typings };
 
         const endClock = startClock();
